+++ InstantiateComponents/Assets/InstantiateComponents/CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
+## [1.0.2] - 2025-01-15
+### Changed
+- Changed to ensure that instances are not regenerated even when component values are changed at runtime.
+- Changed to regenerate the instance when position or rotation changes.
+
 ## [1.0.1] - 2024-12-29 - In Development
 ### Modified
 - Modified the name of the parent object of the instance.
 - Modified the inspector display.
-### Changed
-- Changed to ensure that instances are not regenerated even when component values are changed at runtime.
-- Changed to regenerate the instance when position or rotation changes.
 
 ## [1.0.0] - 2024-12-26
 ### Added